Split out the digest generation from the digest sender
